Change the message logging API.
liblustre doesn't provide an API to log messages. It is the job of the
application to emit these logs. By default no logs are evaluated. To
retrieve the log messages, the application has to set a callback with
llapi_msg_callback_set() and set a log level with
llapi_msg_set_level(). When the callback is called, the application
can then emit the messages if it wishes. lhsmtool_posix.c has been
modified to use that interface. The posix copytool will output all its
messages on stdout.