Changed llapi_hsm_import to use a layout record
Use the new layout to HSM import a file instead of letting
llapi_hsm_import do it. That way the caller can exactly control the
layout.

The function now returns a file descriptor that can be used by the
caller to restore some extended attribute without having to re-open
the file.

The fid is no longer a returned parameter. If the caller wants it, it
can call llapi_fd2fid itself.