Use LL_IOC_LOV_SETSTRIPE instead of setting lustre.lov xattr
When importing a file, directly writing the lustre.lov extended
attribute doesn't work well. Files are created with a stat' block of
0. Their hsm_state is correct (released) but reading from them will
not start an HSM restore operation. Instead it is treated as a sparse
file, leading to data corruption. If restoration is forced by an lfs
hsm_restore, then all is good.